New pipes backend and fixed inframe
Now uses pipes preferentially. Note: many features are not compatible yet. These
features are marked with a LEGACY ONLY tag in the documentation. Use the legacy
argument to use the old fashion method. Also, fixed in frame as it was looking
for gaps OOF which is not of interest now.